requirements.txt
requests~=2.30.0
py2app
charset-normalizer
chardet
python-dotenv
urllib3==1.26.6
terminaltables